move fetch utility to lib/util.js
