+++ lib/util.js
@@ -16,6 +16,44 @@ import { finished } from 'stream/promises'
 const DynamoDBTableName = 'Extensions'
 const FirstVersion = '1.0.0'
 
+/**
+ * Returns a promise that resolves with the URL's body text data.
+ * The request will be retried several times if necessary.
+ *
+ * @param listURL The URL of the list to fetch
+ * @return a promise that resolves with the content of the list or rejects with an error message.
+ */
+const fetchTextFromURL = (listURL) => {
+  const attempt = () => fetch(listURL).then(response => {
+    if (response.status !== 200) {
+      throw new Error(`Error status ${response.status} ${response.statusText} returned for URL: ${listURL}`)
+    }
+    return response.text()
+  }).catch(error => {
+    throw new Error(`Error when fetching ${listURL}: ${error.message}`)
+  })
+
+  // Introduces a delayed rejection into a promise chain so that request retries can be caught in a loop until a request succeeds
+  const delayReject = (ms) => {
+    return (reason) => new Promise((_resolve, reject) => setTimeout(() => {
+      console.log(`Spurious error: ${reason.message}`)
+      console.log('  (retrying...)')
+      reject(reason)
+    }, ms))
+  }
+
+  const maxAttempts = 5
+  const delayMs = 3000
+
+  let p = Promise.reject(new Error())
+
+  for (let i = 0; i < maxAttempts; i++) {
+    p = p.catch(attempt).catch(delayReject(delayMs))
+  }
+
+  return p
+}
+
 const downloadExtensionFromCWS = async (componentId, chromiumVersion, outputPath) => {
   const url = `https://clients2.google.com/service/update2/crx?response=redirect&prodversion=${chromiumVersion}&acceptformat=crx3&x=id%3D${componentId}%26uc`
   const response = await fetch(url)
@@ -360,6 +398,7 @@ const addCommonScriptOptions = (command) => {
 }
 
 export default {
+  fetchTextFromURL,
   createTableIfNotExists,
   downloadExtensionFromCWS,
   generateCRXFile,
